renamed file `idle_manager` to `idle_module`
