+++ pype/modules/idle_manager/idle_module.py
